requirements.txt
-i https://pypi.org/simple
amqp==5.2.0; python_version >= '3.6'
asgiref==3.8.1; python_version >= '3.8'
asttokens==2.4.1
async-timeout==4.0.3; python_full_version < '3.11.3'
billiard==4.2.0; python_version >= '3.7'
celery==5.4.0; python_version >= '3.8'
certifi==2024.7.4; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
cryptography==42.0.8; python_version >= '3.7'
decorator==5.1.1; python_version >= '3.5'
django==4.2.9; python_version >= '3.8'
django-admin-berry==1.0.10
django-filter==24.2; python_version >= '3.8'
django-formset-js==0.5.0
django-guardian==2.4.0; python_version >= '3.5'
django-jquery-js==3.1.1
django-redis==5.4.0; python_version >= '3.6'
djangorestframework==3.15.2; python_version >= '3.8'
exceptiongroup==1.2.2; python_version < '3.11'
executing==2.0.1; python_version >= '3.5'
gunicorn==21.2.0; python_version >= '3.5'
idna==3.7; python_version >= '3.5'
ipython==8.26.0; python_version >= '3.10'
jedi==0.19.1; python_version >= '3.6'
kombu==5.3.7; python_version >= '3.8'
markdown==3.6; python_version >= '3.8'
matplotlib-inline==0.1.7; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
parso==0.8.4; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32' and sys_platform != 'emscripten'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
psycopg2-binary==2.9.9; python_version >= '3.7'
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pyyaml==6.0.1; python_version >= '3.6'
redis==5.0.7; python_version >= '3.7'
requests==2.32.3; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.1; python_version >= '3.8'
stack-data==0.6.3
str2bool==1.1
traitlets==5.14.3; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.12'
tzdata==2024.1; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.2; python_version >= '3.8'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
whitenoise==6.6.0; python_version >= '3.8'
